requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
gunicorn==20.1.0
sqlalchemy==2.0.36
pydantic==2.9.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.0.1
pyjwt==2.8.0
python-multipart==0.0.6
